requirements.txt
praw==2.1.18
APScheduler==3.0.0
PyYAML==3.11